% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limorhyde.R
\name{getCosinorBasis}
\alias{getCosinorBasis}
\title{Basis matrix for cosinor}
\usage{
getCosinorBasis(x, period, intercept)
}
\arguments{
\item{x}{Values of the predictor variable.}

\item{period}{Period for the predictor variable.}

\item{intercept}{If \code{TRUE}, a column of ones will be included in the basis.}
}
\value{
A matrix with a row for each value of \code{x} and a column for each
component of the decomposition.
}
\description{
Generate basis matrix for cosinor regression.
}
\examples{
b = getCosinorBasis(seq(0, 20, 4), period = 24, intercept = FALSE)

}
