% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limorhyde.R
\name{limorhyde}
\alias{limorhyde}
\title{Convert a periodic time variable into components usable in linear models}
\usage{
limorhyde(df, timeColname, period = 24, sinusoid = TRUE, nKnots = 3)
}
\arguments{
\item{df}{Data frame containing one row for each sample.}

\item{timeColname}{Character string indicating the column in \code{df} that contains the time
at which each sample was acquired.}

\item{period}{Number corresponding to the period to use for the decomposition
(in units of the values in \code{timeColname}).}

\item{sinusoid}{If TRUE, a sinusoidal decomposition (sine and cosine) is performed.
If FALSE, a periodic spline decomposition is performed.}

\item{nKnots}{Number of knots in the periodic spline decomposition. Only used if \code{sinusoid}
is FALSE.}
}
\value{
A list.
\item{df}{Data frame, consisting of the input data frame and the new columns for the time
decomposition, that can be used to construct a design matrix for differential rhythmicity analysis.}
\item{timeColnames}{Character vector containing the colnames for the time decomposition.}
}
\description{
\code{limorhyde} decomposes a periodic time variable into multiple components, which enables
detection of differential rhythmicity using state-of-the-art methods for differential expression.
}
\examples{
# create an example data frame
nSamples = 12
df = data.frame(
   sample = paste0('sample_', 1:nSamples),
   genotype = factor(rep(c('WT', 'KO'), each=nSamples/2), levels=c('WT', 'KO')),
   zt = rep(seq(0, 24-24/nSamples*2, 24/nSamples*2), times=2),
   stringsAsFactors=FALSE)

# call limorhyde and examine its output
limoResult = limorhyde(df, 'zt')
limoResult$df
limoResult$timeColnames

# create a design matrix for differential rhythmicity analysis
design = model.matrix(~ genotype * (zt_cos + zt_sin), data=limoResult$df)
}
